Remove dependency on getFilesystemPath
