+++ lib/carbon/tests/test_storage.py
@@ -2,8 +2,6 @@ import os
 from unittest import TestCase
 from mock import patch
 
-from carbon.tests.util import TestSettings
-from carbon.database import WhisperDatabase
 
 # class NoConfigSchemaLoadingTest(TestCase):
 
@@ -85,23 +83,3 @@ class ExistingConfigSchemaLoadingTest(TestCase):
         schema_list = loadAggregationSchemas()
         last_schema = schema_list[-1]
         self.assertEquals(last_schema, defaultAggregation)
-
-
-class getFilesystemPathTest(TestCase):
-
-    def setUp(self):
-        self._sep_patch = patch.object(os.path, 'sep', "/")
-        self._sep_patch.start()
-        settings = TestSettings()
-        settings['LOCAL_DATA_DIR'] = '/tmp/'
-        self._database_patch = patch('carbon.state.database', new=WhisperDatabase(settings))
-        self._database_patch.start()
-
-    def tearDown(self):
-        self._database_patch.stop()
-        self._sep_patch.stop()
-
-    def test_getFilesystemPath(self):
-        from carbon.storage import getFilesystemPath
-        result = getFilesystemPath('stats.example.counts')
-        self.assertEquals(result, '/tmp/stats/example/counts.wsp')
